+++ jjunji/StudyApplication/app/src/main/java/com/example/studyapplication/main/image/ImageFragment.kt
@@ -1,4 +1,4 @@
-package com.example.studyapplication.kin
+package com.example.studyapplication.main.image
 
 import android.content.Context
 import android.os.Bundle
@@ -8,25 +8,25 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.studyapplication.R
-import com.example.studyapplication.kin.adapter.KinAdapter
+import com.example.studyapplication.main.image.adapter.ImageAdapter
 import com.example.studyapplication.network.ApiClient
 import com.example.studyapplication.network.IConn
 import com.example.studyapplication.network.Remote
-import com.example.studyapplication.vo.KinList
-import kotlinx.android.synthetic.main.fragment_kin.*
-import kotlinx.android.synthetic.main.fragment_kin.view.*
+import com.example.studyapplication.vo.ImageList
+import kotlinx.android.synthetic.main.fragment_image.*
+import kotlinx.android.synthetic.main.fragment_movie.view.*
 
-class KinFragment  : Fragment() {
+class ImageFragment : Fragment() {
     lateinit var mContext : Context
 
     companion object {
-        fun newInstance() : KinFragment {
-            return KinFragment()
+        fun newInstance() : ImageFragment {
+            return ImageFragment()
         }
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view : View = inflater.inflate(R.layout.fragment_kin, container, false)
+        val view : View = inflater.inflate(R.layout.fragment_image, container, false)
         view.btnSearch.setOnClickListener(btnSearchClickListener())
         mContext = view.context
 
@@ -36,17 +36,17 @@ class KinFragment  : Fragment() {
     // 검색 버튼 클릭 리스너
     private fun btnSearchClickListener() : View.OnClickListener {
         return View.OnClickListener {
-            val kinTitle = etQuery.text.toString()
-            requestSearchKin(kinTitle)
+            val blogTitle = etQuery.text.toString()
+            requestSearchBlog(blogTitle)
         }
     }
 
-    // 지식인 검색 요청
-    private fun requestSearchKin(title : String) {
-        Remote.get(ApiClient.getService().getKinList(title), object : IConn {
+    // 이미지 검색 요청
+    private fun requestSearchBlog(title : String) {
+        Remote.get(ApiClient.getService().getImageList(title), object : IConn {
             override fun <T> success(result: T) {
-                val kinList : KinList? = result as KinList
-                kinList?.let {
+                val imageList : ImageList? = result as ImageList
+                imageList?.let {
                     setAdapter(it)
                 }
             }
@@ -57,8 +57,8 @@ class KinFragment  : Fragment() {
     }
 
     // recyclerView 세팅
-    private fun setAdapter(response : KinList) {
-        recyclerView.adapter = KinAdapter(response.arrKinInfo)
+    private fun setAdapter(response : ImageList) {
+        recyclerView.adapter = ImageAdapter(response.arrImageInfo)
         recyclerView.layoutManager = LinearLayoutManager(mContext)
     }
 }
