+++ jjunji/StudyApplication/app/src/main/java/com/example/studyapplication/main/image/adapter/ImageHolder.kt
@@ -1,4 +1,4 @@
-package com.example.studyapplication.image.adapter
+package com.example.studyapplication.main.image.adapter
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
