+++ jjunji/StudyApplication/app/src/main/java/com/example/studyapplication/main/kin/KinFragment.kt
@@ -1,4 +1,4 @@
-package com.example.studyapplication.movie
+package com.example.studyapplication.main.kin
 
 import android.content.Context
 import android.os.Bundle
@@ -8,25 +8,25 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.studyapplication.R
-import com.example.studyapplication.movie.adapter.MovieAdapter
+import com.example.studyapplication.main.kin.adapter.KinAdapter
 import com.example.studyapplication.network.ApiClient
 import com.example.studyapplication.network.IConn
 import com.example.studyapplication.network.Remote
-import com.example.studyapplication.vo.MovieList
-import kotlinx.android.synthetic.main.fragment_movie.*
-import kotlinx.android.synthetic.main.fragment_movie.view.*
+import com.example.studyapplication.vo.KinList
+import kotlinx.android.synthetic.main.fragment_kin.*
+import kotlinx.android.synthetic.main.fragment_kin.view.*
 
-class MovieFragment : Fragment() {
+class KinFragment  : Fragment() {
     lateinit var mContext : Context
 
     companion object {
-        fun newInstance() : MovieFragment {
-            return MovieFragment()
+        fun newInstance() : KinFragment {
+            return KinFragment()
         }
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view : View = inflater.inflate(R.layout.fragment_movie, container, false)
+        val view : View = inflater.inflate(R.layout.fragment_kin, container, false)
         view.btnSearch.setOnClickListener(btnSearchClickListener())
         mContext = view.context
 
@@ -36,17 +36,17 @@ class MovieFragment : Fragment() {
     // 검색 버튼 클릭 리스너
     private fun btnSearchClickListener() : View.OnClickListener {
         return View.OnClickListener {
-            val movieTitle = etQuery.text.toString()
-            requestSearchMovie(movieTitle)
+            val kinTitle = etQuery.text.toString()
+            requestSearchKin(kinTitle)
         }
     }
 
-    // 영화 검색 요청
-    private fun requestSearchMovie(title : String) {
-        Remote.get(ApiClient.getService().getMovieList(title), object : IConn {
+    // 지식인 검색 요청
+    private fun requestSearchKin(title : String) {
+        Remote.get(ApiClient.getService().getKinList(title), object : IConn {
             override fun <T> success(result: T) {
-                val movieList : MovieList? = result as MovieList
-                movieList?.let {
+                val kinList : KinList? = result as KinList
+                kinList?.let {
                     setAdapter(it)
                 }
             }
@@ -57,8 +57,8 @@ class MovieFragment : Fragment() {
     }
 
     // recyclerView 세팅
-    private fun setAdapter(response : MovieList) {
-        recyclerView.adapter = MovieAdapter(response.arrMovieInfo)
+    private fun setAdapter(response : KinList) {
+        recyclerView.adapter = KinAdapter(response.arrKinInfo)
         recyclerView.layoutManager = LinearLayoutManager(mContext)
     }
 }
