+++ jjunji/StudyApplication/app/src/main/java/com/example/studyapplication/main/kin/adapter/KinHolder.kt
@@ -1,4 +1,4 @@
-package com.example.studyapplication.kin.adapter
+package com.example.studyapplication.main.kin.adapter
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
