+++ jjunji/StudyApplication/app/src/main/java/com/example/studyapplication/main/movie/MovieFragment.kt
@@ -1,7 +1,6 @@
-package com.example.studyapplication.image
+package com.example.studyapplication.main.movie
 
 import android.content.Context
-import android.media.Image
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,25 +8,26 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.studyapplication.R
-import com.example.studyapplication.image.adapter.ImageAdapter
+import com.example.studyapplication.main.movie.adapter.MovieAdapter
 import com.example.studyapplication.network.ApiClient
 import com.example.studyapplication.network.IConn
 import com.example.studyapplication.network.Remote
-import com.example.studyapplication.vo.ImageList
-import kotlinx.android.synthetic.main.fragment_image.*
+import com.example.studyapplication.vo.MovieList
+import kotlinx.android.synthetic.main.fragment_movie.*
 import kotlinx.android.synthetic.main.fragment_movie.view.*
 
-class ImageFragment : Fragment() {
+class MovieFragment : Fragment() {
     lateinit var mContext : Context
 
     companion object {
-        fun newInstance() : ImageFragment {
-            return ImageFragment()
+        fun newInstance() : MovieFragment {
+            return MovieFragment()
         }
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view : View = inflater.inflate(R.layout.fragment_image, container, false)
+        val view = inflater.inflate(R.layout.fragment_movie, container, false)
+
         view.btnSearch.setOnClickListener(btnSearchClickListener())
         mContext = view.context
 
@@ -37,17 +37,17 @@ class ImageFragment : Fragment() {
     // 검색 버튼 클릭 리스너
     private fun btnSearchClickListener() : View.OnClickListener {
         return View.OnClickListener {
-            val blogTitle = etQuery.text.toString()
-            requestSearchBlog(blogTitle)
+            val movieTitle = etQuery.text.toString()
+            requestSearchMovie(movieTitle)
         }
     }
 
-    // 이미지 검색 요청
-    private fun requestSearchBlog(title : String) {
-        Remote.get(ApiClient.getService().getImageList(title), object : IConn {
+    // 영화 검색 요청
+    private fun requestSearchMovie(title : String) {
+        Remote.get(ApiClient.getService().getMovieList(title), object : IConn {
             override fun <T> success(result: T) {
-                val imageList : ImageList? = result as ImageList
-                imageList?.let {
+                val movieList : MovieList? = result as MovieList
+                movieList?.let {
                     setAdapter(it)
                 }
             }
@@ -58,8 +58,8 @@ class ImageFragment : Fragment() {
     }
 
     // recyclerView 세팅
-    private fun setAdapter(response : ImageList) {
-        recyclerView.adapter = ImageAdapter(response.arrImageInfo)
+    private fun setAdapter(response : MovieList) {
+        recyclerView.adapter = MovieAdapter(response.arrMovieInfo)
         recyclerView.layoutManager = LinearLayoutManager(mContext)
     }
 }
